# Use `--comments` when calling MySQL Client to avoid stripping of TiDB comments