Without this this happens:
```
mysql> SELECT /*T 'ti', */ 'db';
+----+
| db |
+----+
| db |
+----+
1 row in set (0.00 sec)
```

The `/*T ... */` syntax is use in the output of `SHOW CREATE TABLE...`
and other locations.