@baurine 2021-08-19
hi @dveeden , I found it has the same output for `SELECT /*T 'ti', */ 'db';` between with `--comments` and without `--comments`, can you add some more explanation or ref docs for this option, thanks!

btw, I found in the output of `tiup playground` command, it says `To connect TiDB: mysql --host 127.0.0.1 --port 4000 -u root -p (no password) --comments`.