@dveeden 2021-08-19
@baurine sorry, my mistake. It is `/*T!` instead of `/*T`:
```
[dvaneeden@dve-carbon ~]$ mysql -h 127.0.0.1 -u root -P 4000 --comments -Ne "SELECT /*T! 'ti', */ 'db'"
+----+----+
| ti | db |
+----+----+
[dvaneeden@dve-carbon ~]$ mysql -h 127.0.0.1 -u root -P 4000 -Ne "SELECT /*T! 'ti', */ 'db'"
+----+
| db |
+----+
```

Here is an example:

Take a table that is created with clustered index enabled:
```
mysql> create table t (id int primary key);
Query OK, 0 rows affected (0.01 sec)

mysql> show create table t\G
*************************** 1. row ***************************
       Table: t
Create Table: CREATE TABLE `t` (
  `id` int(11) NOT NULL,
  PRIMARY KEY (`id`) /*T![clustered_index] CLUSTERED */
) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin
1 row in set (0.00 sec)
```

Now drop the table and create it again copying the output from the `SHOW CREATE TABLE`:

Without `--comments`:
```
mysql> SET tidb_enable_clustered_index=OFF;
Query OK, 0 rows affected (0.00 sec)

mysql> CREATE TABLE `t` (
    ->   `id` int(11) NOT NULL,
    ->   PRIMARY KEY (`id`) /*T![clustered_index] CLUSTERED */
    -> ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin;
Query OK, 0 rows affected (0.00 sec)

mysql> show create table t\G
*************************** 1. row ***************************
       Table: t
Create Table: CREATE TABLE `t` (
  `id` int(11) NOT NULL,
  PRIMARY KEY (`id`) /*T![clustered_index] NONCLUSTERED */
) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin
1 row in set (0.00 sec)
```

With `--comments`:
```
mysql> SET tidb_enable_clustered_index=OFF;
Query OK, 0 rows affected (0.00 sec)

mysql> CREATE TABLE `t` (
    ->   `id` int(11) NOT NULL,
    ->   PRIMARY KEY (`id`) /*T![clustered_index] CLUSTERED */
    -> ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin;
Query OK, 0 rows affected (0.00 sec)

mysql> show create table t\G
*************************** 1. row ***************************
       Table: t
Create Table: CREATE TABLE `t` (
  `id` int(11) NOT NULL,
  PRIMARY KEY (`id`) /*T![clustered_index] CLUSTERED */
) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin
1 row in set (0.00 sec)
```

The docs:
https://dev.mysql.com/doc/refman/8.0/en/mysql-command-options.html#option_mysql_comments

The code:
https://github.com/mysql/mysql-server/blob/beb865a960b9a8a16cf999c323e46c5b0c67f21f/client/mysql.cc#L2447-L2670

MySQL has whitelisted some comments: `/*!` (version specific comment) and `/*+` (optimizer hints) but not `/*T`